% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwselector.R
\name{bwselector}
\alias{bwselector}
\title{Selector of window}
\usage{
bwselector(
  muestra,
  dim = ncol(muestra) - 1,
  X = muestra[, 1:dim],
  Y = muestra[, dim + 1],
  malla = mallador(X, Y, dim = dim, x.malla = X),
  eps = 1e-08,
  p = floor(-log(eps, base = 10)),
  nivel = 0.95,
  metodo = "CV",
  n = length(Y),
  eps2 = 0.001
)
}
\arguments{
\item{eps}{para la convergencia del PMS}

\item{eps2}{para el criterio de parada del algoritmo heurístico (que parará cuando la diferencia de los tres puntos sea pequeña, reescalando por el mayor)}
}
\description{
Selector of window
}
