% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVC.R
\name{CVC}
\alias{CVC}
\title{Compute the cross-validation like function suggested by Zhou and Huang}
\usage{
CVC(
  data = cbind(X, Y),
  X = data[, 1:dim],
  Y = data[, dim + 1],
  malla = mallador(data),
  dim,
  h1,
  h2,
  p = floor(-log(eps, base = 10)),
  eps = 10^(-p),
  n = length(Y),
  k = attr(malla, "k")
)
}
\arguments{
\item{data}{Matrix or data frame containing the sample.}

\item{X}{Matrix with the covariable points.}

\item{Y}{Vector with the response values.}

\item{malla}{An \code{MRpms_malla} object returned by the \code{mallador} function with the initial guesses to comppute the modes.}

\item{dim}{Dimension of covariable.}

\item{h1}{Smooth parameter for covariable.}

\item{h2}{Smooth parameter for response.}

\item{p}{Decimal precision.}

\item{eps}{Convergence tolerance.}

\item{n}{Sample size.}

\item{k}{Number of guesses for each covariable point in mesh.}
}
\value{
The value of the cross-validation like function suggested by Zhou and Huang,

\deqn{\displaystyle \frac{1}{n}\sum_{i=1}^n d^2\left(\widehat{M}_{n,\mathbf{h},-i}\left(X_i\right),Y_i\right)\widehat{N}_{\mathbf{h},-i}^2\left(X_i\right)w\left(X_i\right),}

where \eqn{\widehat M_{n,\mathbf{h},-i}(X_i)} is the estimated modes in \eqn{X_i} without consider \eqn{X_i} and \eqn{\widehat{N}_{\mathbf{h},-i}^2\left(X_i\right)w\left(X_i\right)}
is the number of modes estimated in \eqn{X_i}.
}
\description{
This function mask a C function which preforms the cross validation
suggested by Zhou and Huang to search the optimal smooth parameter.
}
\references{
Zhou, H. and Huang, X. (2019).
\emph{Bandwidth selection for nonparametric modal regression}.
Communication in Statistics - Simulation and Computation,
\strong{48}(4), 968--984.
}
\keyword{internal}
