% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Ejemplo1}
\alias{Ejemplo1}
\title{Ejemplo 1}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 200 rows and 2 columns.
}
\usage{
Ejemplo1
}
\description{
Dos senos a distintas alturas
}
\examples{
plot(Ejemplo1)

}
\keyword{datasets}
