% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hP.R
\name{hP}
\alias{hP}
\title{Compute the volume of the uniform prediction set.}
\usage{
hP(
  X,
  Y,
  malla,
  dim = 1,
  h1,
  h2,
  p = floor(-log(eps, base = 10)),
  eps = 10^(-p),
  n = length(Y),
  conf.level = 0.95,
  k = attr(malla, "k"),
  lensopx = max(X) - min(X)
)
}
\arguments{
\item{X}{Matrix with the coordinates of the covariable vector.}

\item{Y}{Vector with the response variable values.}

\item{malla}{A \code{malla} object from the \code{mallador} function.}

\item{dim}{Dimension of covariable vector.}

\item{h1}{Smooth parameter for covariable.}

\item{h2}{Smooth parameter for response variable.}

\item{p}{Number of decimal precision.}

\item{eps}{Tolerance for convergence in the partial Mean-Shift.}

\item{n}{Sample size.}

\item{conf.level}{Confidence level for the uniform prediction set.}

\item{k}{Number of initial guesses in every covariable point in the mesh.}

\item{lensopx}{Size of the support of covariable.}
}
\value{
The volume of the estimated uniform prediction set.
}
\description{
Compute the volume of the uniform prediction set for a fitted model. First compute the modes in a given mesh
using cross-validation, compute the quantile for a given confidence level and later use Monte-Carlo integration. Currently,
only works in the univariate case.
}
