% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mallador.R
\name{mallador}
\alias{mallador}
\title{Construct a mesh compatible with MRpms functions.}
\usage{
mallador(
  X = muestra[, 1:dim],
  Y = muestra[, dim + 1],
  k = 10,
  len = 100,
  muestra = cbind(X, Y),
  dim = 1,
  x.malla
)
}
\arguments{
\item{X}{Matrix with covariable data points.}

\item{Y}{Vector with response variable data values.}

\item{k}{Number of \code{y} values for each \code{x} point in the new mesh.}

\item{len}{Number of diferent \code{x} points in the new mesh.}

\item{muestra}{Matrix with data points.}

\item{dim}{Dimension of covariable.}

\item{x.malla}{Matrix of \code{x} points given by user.}
}
\value{
A matrix with the new mesh. Each \code{x} point is repeated \code{k} times, each one with
their associated \code{y} value. Besides, the matrix have three attributes: \code{l}, \code{k} and \code{x.malla}
which contains a matrix with the diferents \code{x} points.
}
\description{
Given a matrix with data points, or two matrix with the covariable points and response values,
construct a mesh of points in both ways: compute ´l´ diferent equidistant \code{x} points and asign \code{k} diferent \code{y}
values to each ones between the nearest data points; or take a given \code{x} matrix of points and preforms the same construction.
}
