% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRpms_conf.R
\name{plot.MRpms_conf}
\alias{plot.MRpms_conf}
\title{Plot method for \code{MRpms_conf} class with dimension \code{1}.}
\usage{
\method{plot}{MRpms_conf}(x, data, ...)
}
\arguments{
\item{x}{An \code{MRpms_conf} object.}

\item{data}{The data from which the modes were computed.}

\item{...}{Other parameters}
}
\value{
A plot with the \code{data}, in blank circles, the modes in red, the pointwise confidence intervals in blue
and the uniform confidence intervals in gray.
}
\description{
Plot method for \code{MRpms_conf} objects, which
are computed from a \code{ConfPMS} call and contains information to construct confidence intervals
for the computed modes.
}
\seealso{
\code{\link[=ConfPMS]{ConfPMS()}}
}
