% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRpms_modas.R
\name{plot.MRpms_modas}
\alias{plot.MRpms_modas}
\title{Plot method for \code{MRpms_modas} class with dimension \code{1}.}
\usage{
\method{plot}{MRpms_modas}(x, data, ...)
}
\arguments{
\item{x}{An \code{MRpms_modas} object to plot.}

\item{data}{The data from which the modes were computed.}

\item{...}{Other parameters}
}
\value{
A plot with both the \code{data} and the modes.
}
\description{
Plot method for \code{MRpms_modas} objects, which
are computed from a \code{PMS} call or, a more bare-bones version, \code{PMSc} call.
}
\seealso{
\code{\link[=PMS]{PMS()}}
}
