% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRpms_pred.R
\name{plot.MRpms_pred}
\alias{plot.MRpms_pred}
\title{Plot method for \code{MRpms_pred} class with dimension \code{1}.}
\usage{
\method{plot}{MRpms_pred}(x, data, ...)
}
\arguments{
\item{x}{An \code{MRpms_pred} object.}

\item{data}{The data from which the modes were computed.}

\item{...}{Other parameters}
}
\value{
A plot with the \code{data}, in blank circles and the uniform prediction intervals in gray.
}
\description{
Plot method for \code{MRpms_pred} objects, which
are computed from a \code{PredPMS} call and contains information to construct prediction intervals
for the computed modes.
}
\seealso{
\code{\link[=PredPMS]{PredPMS()}}
}
