% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMSr.R
\name{PMS}
\alias{PMS}
\title{Performs modal regression}
\usage{
PMS(
  muestra,
  x.malla = NULL,
  h1 = 0.5,
  h2 = 1,
  eps = 1e-08,
  p = floor(-log(eps, base = 10)),
  k = 10,
  l = 200,
  malla = NULL
)
}
\arguments{
\item{muestra}{Matrix containing the sample.}

\item{x.malla}{Set of covariate points where the modes will be estimated.}

\item{h1}{Smoothing parameter for covariates.}

\item{h2}{Smoothing parameter for response variable.}

\item{eps}{Convergence tolerance and threshold to discriminate between modes.}

\item{p}{Number of decimal digits considered when discriminating modes.}

\item{k}{Number of Y values per x point in the \code{malla} object created.}

\item{l}{Number of different X points in the \code{malla} object.}

\item{malla}{Matrix containing the initial points used to compute the modes
with the Partial Mean-Shift algorithm. If not provided,
\code{mallador}function is called. It must be a MRpms_malla class object.}
}
\value{
A list containing the estimated local modes conditional on the values
of the covariate appearing in \code{x.malla}. In addition, \code{x.malla} is returned as an attribute.
}
\description{
Given sample with one-dimensional covariate, performs modal regression
on a given set of covariate values.
}
