% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConjuntoPredUniforme.R
\name{PredPMS}
\alias{PredPMS}
\title{Uniform prediction sets based on modal regression.}
\usage{
PredPMS(
  muestra,
  modas,
  x.malla,
  h1 = 0.4,
  h2 = 1,
  eps = 1e-08,
  p = -log(eps, base = 10),
  conf.level = 0.95,
  k = 10,
  l = 200,
  malla
)
}
\arguments{
\item{muestra}{Matrix containing the sample.}

\item{modas}{List containing the estimated set of conditional local
modes on each covariate point in \code{malla}. In case modas missing,
PMSc is called to compute them. It must be a MRpms_modas class object.}

\item{x.malla}{Set of covariate points where the modes will be estimated.
Not needed if \code{modas} is provided.}

\item{h1}{Smoothing parameter for covariates.}

\item{h2}{Smoothing parameter for response variable.}

\item{eps}{Convergence tolerance and threshold to discriminate between modes.}

\item{p}{Number of decimal digits considered when discriminating modes.}

\item{conf.level}{Confidence level for the prediction set.}

\item{k}{Number of Y values per \code{x} point in the \code{malla} object created,
if \code{malla} is not provided.}

\item{l}{Number of different \code{x} points in the \code{malla} object.}

\item{malla}{Matrix containing the initial points used to compute the
modes using the Partial Mean-Shift algorithm. If not provided,
\code{mallador} function is called. It must be a MRpms_malla class object.}
}
\value{
A number, which is the radius of the prediction set,
and a plot of the prediction set, the estimated modes and the sample,
if the covariate is one-dimensional.
}
\description{
Computes a uniform prediction set for the response value.
}
