% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRpms_conf.R
\name{print.MRpms_conf}
\alias{print.MRpms_conf}
\title{Print an \code{MRpms_conf} object}
\usage{
\method{print}{MRpms_conf}(x, ...)
}
\arguments{
\item{x}{An \code{MRpms_conf} object.}

\item{...}{Other arguments. Currently, nothing more is avaliable.}
}
\value{
An index of contents.
}
\description{
A generic function to print an \code{MRpms_conf} object, which are computed
from a \code{ConfPMS} call.
}
