% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRpms_modas.R
\name{print.MRpms_modas}
\alias{print.MRpms_modas}
\title{Print an \code{MRpms_modas} object}
\usage{
\method{print}{MRpms_modas}(x, ...)
}
\arguments{
\item{x}{An \code{MRpms_modas} object.}

\item{...}{Other arguments. Currently, nothing more is avaliable.}
}
\value{
The modes in every covariable point from \code{x.malla}, which is an attribute of the \code{MRpms_modas}.
}
\description{
A generic function to print an \code{MRpms_modas} object, which are computed
from a \code{PMS} cal or, a more bare-bones version, \code{PMSc} call.
}
