% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRpms_pred.R
\name{print.MRpms_pred}
\alias{print.MRpms_pred}
\title{Print an \code{MRpms_pred} object}
\usage{
\method{print}{MRpms_pred}(x, ...)
}
\arguments{
\item{x}{An \code{MRpms_pred} object.}

\item{...}{Other arguments. Currently, nothing more is avaliable.}
}
\value{
The radius of the uniform prediction interval and the first modes.
}
\description{
A generic function to print an \code{MRpms_pred} object, which are computed
from a \code{PredPMS} call.
}
