% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRpms_conf.R
\name{summary.MRpms_conf}
\alias{summary.MRpms_conf}
\title{Print useful information of an \code{MRpms_conf} object}
\usage{
\method{summary}{MRpms_conf}(object, ...)
}
\arguments{
\item{object}{An \code{MRpms_conf} object.}

\item{...}{Other arguments. Currently, nothing more is avaliable.}
}
\value{
The summarize of the \code{MRpms_conf} object, plus estimated uniform error if computed and quantiles of pointwise errors if computed.
}
\description{
A generic function to summarize useful information of an \code{MRpms_conf} object, which are computed
from a \code{ConfPMS} call.
}
