% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRpms_modas.R
\name{summary.MRpms_modas}
\alias{summary.MRpms_modas}
\title{Print useful information of an \code{MRpms_modas} object}
\usage{
\method{summary}{MRpms_modas}(object, ...)
}
\arguments{
\item{object}{An \code{MRpms_modas} object.}

\item{...}{Other arguments. Currently, nothing more is avaliable.}
}
\value{
The maximum, minimum and mean number of modes computed, with certain tolerance to avoid atipic values.
}
\description{
A generic function to summarize useful information of an \code{MRpms_modas} object, which are computed
from a \code{PMS} call or, a more bare-bones version, \code{PMSc} call.
}
