% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRpms_pred.R
\name{summary.MRpms_pred}
\alias{summary.MRpms_pred}
\title{Print useful information of an \code{MRpms_pred} object}
\usage{
\method{summary}{MRpms_pred}(object, ...)
}
\arguments{
\item{object}{An \code{MRpms_pred} object.}

\item{...}{Other arguments. Currently, nothing more is avaliable.}
}
\value{
The summarize of the \code{MRpms_modas} object, plus estimated radius for uniform prediction set and volume of said region.
}
\description{
A generic function to summarize useful information of an \code{MRpms_pred} object, which are computed
from a \code{PredPMS} call.
}
