% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{twosines}
\alias{twosines}
\title{Two sines shifted in Y axis.}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 200 rows and 2 columns.
}
\usage{
twosines
}
\description{
Two sines shifted in Y axis.
}
\details{
The 200 points from two sines with a certain normal error,
\itemize{
\item 100 from the usual sine ploted between 0 and 10 (in radians),
\item 100 from a shifted sine 3 points lower between the same support.
}

In both cases, error follows a certain \eqn{\mathcal N(0,0.3)}.
}
\examples{
plot(twosines)

}
\keyword{datasets}
