% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwselector.R
\name{bwselector}
\alias{bwselector}
\title{Bandwith selector}
\usage{
bwselector(
  muestra,
  malla = mallador(muestra, x.malla = X),
  eps = 1e-08,
  conf.level = 0.95,
  metodo = "CV",
  eps2 = 0.001
)
}
\arguments{
\item{muestra}{Matrix containing the sample.}

\item{malla}{Matrix containing the initial points used to compute the modes
using the Partial Mean-Shift algorithm. If not provided,
\code{mallador}function is called. It must be a MRpms_malla class object.}

\item{eps}{PMS convergence tolerance.}

\item{conf.level}{Confidence level for prediction sets (if \code{metodo}=\code{P}).}

\item{metodo}{\code{CV} (default) for Zhao and Huang selector or \code{P} for
Chen et al. selector based on the size of prediction sets.}

\item{eps2}{Convergence tolerance for heuristic algorithm.}
}
\value{
A vector with two elements, \code{h1} and \code{h2}.
}
\description{
Performs an heuristic search based on the Nedler-Mead algorithm to find
appropiate values for the bandwith parameters \code{h1} and \code{h2}.
(Warning: this is a computationally requiring method which
might take some time to finish).
}
\examples{
h <- bwselector(twosines)
modas <- PMS(twosines, h1 = h[1], h2 = h[2])

plot(twosines)
plot(modas, pch = 19, col = "red")

}
\references{
Chen, Y.-C., Genovese, C. R., Tibshirani, R. J. and Wasserman, L. (2016).
\emph{Nonparametric modal regression}. The Annals of Statistics, \strong{44}(2), 489--514.

Zhou, H. and Huang, X. (2019).
\emph{Bandwidth selection for nonparametric modal regression}.
Communication in Statistics - Simulation and Computation,
\strong{48}(4), 968--984.
}
