% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConfPMS.R
\name{ConfPMS}
\alias{ConfPMS}
\title{Pointwise and uniform confidence sets}
\usage{
ConfPMS(
  muestra,
  modas,
  malla,
  h1 = 0.3,
  h2 = 0.5,
  eps = 1e-08,
  k = 10,
  len = 200,
  conf.level = 0.95,
  B = 30,
  type = 2,
  seed = 2026
)
}
\arguments{
\item{muestra}{Matrix containing the sample.}

\item{modas}{List containing the estimated set of conditional local modes
on each covariate point in \code{malla}. In case \code{modas} missing,
\code{PMSc}is called to compute them. It must be a MRpms_modas class object.}

\item{malla}{Matrix containing the initial points used to compute the modes
using the Partial Mean-Shift algorithm. If not provided,
\code{mallador}function is called. It must be a MRpms_malla class object.}

\item{h1}{Smoothing parameter for covariates.}

\item{h2}{Smoothing parameter for response variable.}

\item{eps}{PMS convergence tolerance.}

\item{k}{Number of Y values per x point in the \code{malla} object created.}

\item{len}{Number of different X points in the \code{malla} object.}

\item{conf.level}{Confidence level for the sets.}

\item{B}{Number of \emph{bootstrap} replications computed to estimate
the confidence sets.}

\item{type}{Numerical argument. "0" means computing pointwise sets only,
"1" for uniform sets only and "2" for both. Set to "2" by default.}

\item{seed}{Seed for the \emph{bootstrap} simulations.}
}
\value{
A list. First entry is a list containing the estimated set of
conditional local modes for each \code{x} point in the \code{malla}. If \code{type} is 0 or 2,
returns a vector with the estimated pointwise errors (\code{deltax}).
If \code{type} is 1 or 2, returns the estimated uniform error (\code{delta}).
}
\description{
Computes pointwise and uniform confidence sets for conditional local modes
and plots them along with the estimated modes over the given sample.
(Warning: a large number of bootstrap replications increases execution time).
}
\examples{
conf <- ConfPMS(twosines, B = 10)


}
\references{
Chen, Y.-C., Genovese, C. R., Tibshirani, R. J. and Wasserman, L. (2016).
\emph{Nonparametric modal regression}. The Annals of Statistics, \strong{44}(2), 489--514.
}
