% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mallador.R
\name{mallador}
\alias{mallador}
\title{Constructs a mesh compatible with MRpms functions}
\usage{
mallador(data, k = 10, len = 200, x.malla)
}
\arguments{
\item{data}{Matrix or data frame containing the sample.}

\item{k}{Number of \code{y} values for each \code{x} point in the new mesh.}

\item{len}{Number of different \code{x} points in the new mesh.}

\item{x.malla}{Matrix of \code{x} points given by user.}
}
\value{
A matrix containing the points of the new mesh.
Each \code{x} point is repeated \code{k} times, each one with an associated \code{y} value.
This matrix has three attributes: \code{len}, \code{k} and \code{x.malla}, the last of which
is a matrix containing the different \code{x} points in the mesh.
}
\description{
Given a matrix containing the sample, constructs a mesh.
This can be done in two ways: either computing \code{l} different equidistant \code{x}
points and assigning \code{k} different \code{y} values to each one,
chosen in a k nearest neighbors fashion or, if given a matrix of \code{x}
points (\code{x.malla}), performing the same construction over them.
}
\examples{
malla <- mallador(twosines)
plot(twosines)
points(malla, pch = 19, cex = 0.8)

}
