% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRpms-package.R
\docType{package}
\name{MRpms-package}
\alias{MRpms}
\alias{MRpms-package}
\title{MRpms: Modal Regression with partial mean shift.}
\description{
Preforms unidimensional and multidimensional modal regression based in the Partial Mean-Shift algorithm.
}
\author{
\strong{Maintainer}: Iker iker.rodriguez.rodriguez@rai.usc.es Rodríguez Rodríguez

Authors:
\itemize{
  \item Ángel Antonio angelantonio.perez@rai.usc.es Pérez Balo
}

}
\keyword{internal}
