% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRpms-package.R
\docType{package}
\name{MRpms-package}
\alias{MRpms}
\alias{MRpms-package}
\title{MRpms: Modal Regression with partial mean shift}
\description{
Performs modal regression based in the Partial Mean-Shift algorithm for one-dimensional covariates.
}
\author{
\strong{Maintainer}: Iker Rodríguez Rodríguez \email{iker.rodriguez.rodriguez@rai.usc.es}

Authors:
\itemize{
  \item Ángel Antonio Pérez Balo \email{angelantonio.perez@rai.usc.es}
}

}
\keyword{internal}
