% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMSr.R
\name{PMS}
\alias{PMS}
\title{Preforms the Partial Mean-Shift Algorithm}
\usage{
PMS(
  data,
  x.malla = NULL,
  h1 = 0.3,
  h2 = 0.5,
  eps = 1e-08,
  k = 10,
  len = 200,
  dim.y = 1,
  malla = NULL
)
}
\arguments{
\item{data}{Matrix or data frame containing the sample. Currently, the last column is needed to be the response values.}

\item{x.malla}{Set of covariate points where the modes will be estimated.}

\item{h1}{Smoothing parameter for covariates.}

\item{h2}{Smoothing parameter for response variable.}

\item{eps}{Convergence tolerance.}

\item{k}{Number of Y values per x point in the \code{malla} object created.}

\item{len}{Number of different X points in the \code{malla} object.}

\item{dim.y}{Response dimension.}

\item{malla}{Matrix containing the initial points used to compute the modes
with the Partial Mean-Shift algorithm. If not provided,
\code{mallador}function is called. It must be a \code{MRpms_malla} class object.}
}
\value{
A \code{MRpms_modas} object, which contains
\itemize{
\item \verb{$modas}: contains the estimated local modes conditional on the values
of the covariate appearing in \code{x.malla};
\item \verb{$x.malla}: with each covariable point where modes were computed;
\item \verb{$dims}: Covariable and response dimensions.
}
}
\description{
Finds the local modes of \code{data} conditional on a given set of covariate values.
}
\examples{
modas <- PMS(twosines)
plot(modas,twosines,pch = 19, col = "red")

}
\references{
Chen, Y.-C., Genovese, C. R., Tibshirani, R. J. and Wasserman, L. (2016).
\emph{Nonparametric modal regression}. The Annals of Statistics, \strong{44}(2), 489--514.
}
