% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMSr.R
\name{PMS}
\alias{PMS}
\title{Performs modal regression over a given sample of unidimensional or multidimensional covariable.}
\usage{
PMS(
  muestra,
  x.malla,
  h1 = 0.5,
  h2 = 1,
  p = floor(-log(eps, base = 10)),
  eps = 10^(-p),
  k,
  l,
  malla
)
}
\arguments{
\item{muestra}{Matrix which contains the data points.}

\item{x.malla}{Set of covariable points where compute the modes.}

\item{h1}{Smoothing parameter for covariables.}

\item{h2}{Smoothing parameter for response variable.}

\item{p}{Number of significance digits considered to diference modes.}

\item{eps}{Tolerance of convergence and minimum diference between modes.}

\item{k}{Number of Y values per x value in the \code{malla} object created, if \code{malla} is not provided.}

\item{l}{Number of X diferent points computed in the \code{malla} object.}

\item{malla}{Matrix of points used to compute the modes using the Partial Mean-Shift algorithm. If not provided, function calls  \code{mallador} to compute.}
}
\description{
Performs modal regression over a given sample of unidimensional or multidimensional covariable.
}
