% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConjuntoPredUniforme.R
\name{PredPMS}
\alias{PredPMS}
\title{Preforms prediction intervals for each mode.}
\usage{
PredPMS(
  muestra,
  modas,
  dim = ncol(muestra) - 1,
  X = muestra[, 1:dim],
  Y = muestra[, dim + 1],
  malla = mallador(X, Y, dim = dim),
  h1 = 0.4,
  h2 = 1,
  n = length(Y),
  eps = 1e-08,
  p = -log(eps, base = 10),
  conf.level = 0.95
)
}
\arguments{
\item{muestra}{Matrix which contains the data points.}

\item{modas}{List of vector which contains the modes on each covariable point of \code{malla}. In case \code{modas} missing, the function calls \code{PMSc} to compute them.}

\item{dim}{Dimension of covariable.}

\item{X}{Matrix of covariable data points in case \code{muestra} not present.}

\item{Y}{Vector with response values in case \code{muestra} not present.}

\item{malla}{Matrix of points used to compute the modes using the Partial Mean-Shift algorithm. If not provided, function calls  \code{mallador} to compute.}

\item{h1}{Smoothing parameter for covariables.}

\item{h2}{Smoothing parameter for response variable.}

\item{eps}{Tolerance of convergence and minimum diference between modes.}

\item{p}{Number of significance digits considered to diference modes.}

\item{conf.level}{Confidence level of intervals.}

\item{type}{Numerical argument. "0" to compute pointwise intervals, "1" to uniform intervals or "2" to compute both. By default, preforms both.}
}
\description{
With this function, the user can preform prediction intervals for response value in each
\code{x} point from a \code{malla} object generated by the \code{mallador} function.
}
