% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConjuntoPredUniforme.R
\name{PredPMS}
\alias{PredPMS}
\title{Preforms prediction intervals for each mode.}
\usage{
PredPMS(
  muestra,
  modas,
  x.malla,
  h1 = 0.4,
  h2 = 1,
  eps = 1e-08,
  p = -log(eps, base = 10),
  conf.level = 0.95,
  k,
  l,
  malla
)
}
\arguments{
\item{muestra}{Matrix which contains the data points.}

\item{modas}{List of vector which contains the modes on each covariable point of \code{malla}. In case \code{modas} missing, the function calls \code{PMSc} to compute them.}

\item{x.malla}{Set of covariable points where compute the modes. Not needed if a \code{modas} object is provided.}

\item{h1}{Smoothing parameter for covariables.}

\item{h2}{Smoothing parameter for response variable.}

\item{eps}{Tolerance of convergence and minimum diference between modes.}

\item{p}{Number of significance digits considered to diference modes.}

\item{conf.level}{Confidence level of intervals.}

\item{malla}{Matrix of points used to compute the modes using the Partial Mean-Shift algorithm. If not provided, function calls  \code{mallador} to compute.}
}
\description{
With this function, the user can preform prediction intervals for response value in each
\code{x} point from a \code{malla} object generated by the \code{mallador} function.
}
