% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredPMS.R
\name{PredPMS}
\alias{PredPMS}
\title{Uniform prediction sets based on modal regression}
\usage{
PredPMS(
  data,
  modas,
  x.malla,
  h1 = 0.3,
  h2 = 0.5,
  dim.y = 1,
  eps = 1e-08,
  conf.level = 0.95,
  k = 10,
  len = 200,
  malla
)
}
\arguments{
\item{data}{Matrix or data frame containing the sample. Currently,
the last column is needed to be de response values.}

\item{modas}{List containing the estimated set of conditional local
modes on each covariate point in \code{malla}. In case modas missing,
PMSc is called to compute them. It must be a \code{MRpms_modas} class object.}

\item{x.malla}{Set of covariate points where the modes will be estimated.
Not needed if \code{modas} is provided.}

\item{h1}{Smoothing parameter for covariates.}

\item{h2}{Smoothing parameter for response variable.}

\item{dim.y}{Response dimension.}

\item{eps}{Convergence tolerance and threshold to discriminate between modes.}

\item{conf.level}{Confidence level for the prediction set.}

\item{k}{Number of Y values per \code{x} point in the \code{malla} object created,
if \code{malla} is not provided.}

\item{len}{Number of different \code{x} points in the \code{malla} object.}

\item{malla}{Matrix containing the initial points used to compute the
modes using the Partial Mean-Shift algorithm. If not provided,
\code{mallador} function is called. It must be a MRpms_malla class object.}
}
\value{
An \code{MRpms_pred} object which contains,
\itemize{
\item \code{modas}: an \code{MRpms_modas} object with the adjusted modes,
\item \code{epsh}: the radius of the uniform prediction set.
\item \code{conf.level}: the confidence level used to compute \code{epsh}.
}

Aditionally, the prediction set and the estimated modes are plotted in case
\code{dim.x} and \code{dim.y} are \code{1}.
}
\description{
Computes a uniform prediction set for the response value.
}
\examples{
pred <- PredPMS(twosines)

}
\references{
Chen, Y.-C., Genovese, C. R., Tibshirani, R. J. and Wasserman, L. (2016).
\emph{Nonparametric modal regression}. The Annals of Statistics, \strong{44}(2), 489--514.
}
